+++ acefile.py
@@ -154,6 +154,12 @@ def c_div(q, d):
     s = int(math.copysign(1, q) * math.copysign(1, d))
     return s * int(abs(q) / abs(d))
 
+def c_char(i):
+    """
+    Convert arbitrary integer to c signed char type range as if casted in c.
+    """
+    return ((i + 128) % 256) - 128
+
 def c_uchar(i):
     """
     Convert arbitrary integer to c unsigned char type range as if casted in c.
@@ -1684,7 +1690,7 @@ class Sound:
         def rar_adjust(self, sample):
             self.__byte_count += 1
             pred_sample = self._get_predicted_sample()
-            pred_dif = (pred_sample - sample) << 3
+            pred_dif = (c_char(pred_sample - sample) << 3)
             self.__rar_dif[0] += abs(pred_dif - self.__rar_dif_cnt[0])
             self.__rar_dif[1] += abs(pred_dif + self.__rar_dif_cnt[0])
             self.__rar_dif[2] += abs(pred_dif - self.__rar_dif_cnt[1])
@@ -1695,9 +1701,9 @@ class Sound:
             self.__rar_dif[7] += abs(pred_dif + self.__rar_dif_cnt[3])
             self.__rar_dif[8] += abs(pred_dif)
 
-            self.__last_delta = sample - self.__last_sample
+            self.__last_delta = c_char(sample - self.__last_sample)
             self.__pred_dif_cnt[0] += self._quantizer[pred_dif >> 3]
-            self.__pred_dif_cnt[1] += self._quantizer[-self.__last_delta]
+            self.__pred_dif_cnt[1] += self._quantizer[self.__last_sample-sample]
             self.__last_sample = sample
 
             if self.__byte_count & 0x1F == 0:
@@ -1776,7 +1782,7 @@ class Sound:
                 break
             sample = c_uchar(value + self.__channels[channel].rar_predict())
             chunk.append(sample)
-            self.__channels[channel].rar_adjust(sample)
+            self.__channels[channel].rar_adjust(c_char(sample))
         return (chunk, next_mode)