+++ NEWS.md
@@ -1,5 +1,6 @@
 ### acefile 0.6.4-dev
 
+-   Fix signedness of ACE 2.0 SOUND mode diff calculations (issue #5).
 -   Add basic debugging facility: `acefile.DEBUG = True` and CLI `--debug`
     hidden option.