+++ NEWS.md
@@ -1,3 +1,6 @@
+### acefile 0.6.5-dev
+
+
 ### acefile 0.6.4 2017-07-26
 
 -   Fix signedness of ACE 2.0 SOUND mode diff calculations (issue #5).